pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "shaarit"
version = "0.0.1"
description = "ShaarIt: Share your thoughts, links, ideas, notes. self-hosting bookmarks and micro blogging. If you ever wanted Shaarli in python, that's it ^_* (w/ starlette)"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "폭스마스크", email = "foxmask+git@pm.me" }]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: WTFPL",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Communications",
    "Topic :: Database",
]

dependencies = [
    "starlette == 0.42.0",
    "starlette-wtf == 0.4.5",
    "uvicorn == 0.34.0",
    "tortoise-orm == 0.23.0",
    "Markdown == 3.7",
    "Pygments == 2.18.0",
    "rich == 13.9.4",
    "python-slugify == 8.0.4",
    "Jinja2 == 3.1.5",
]

[project.urls]
Homepage = "https://github.com/foxmask/shaarit/"
Documentation = "https://github.com/foxmask/shaarit/blob/main/README.md"
Source = "https://github.com/foxmask/shaarit/"

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "F",   # pyflakes
    "DJ",  # flake8 django
    "E",   # pycodestyle
    "I",   # isort
    "ANN", # flake8 type annotations
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.isort]
combine-as-imports = true

[project.optional-dependencies]
test = [
    "build",
    "wheel",
    "coverage",
    "pipdeptree",
    "twine",
    "bandit",
    "pip-check",
    "pip-audit",
]
lint = ["ruff == 0.7.3"]

[tool.setuptools.packages.find]
include = ["shaarit*"]
namespaces = false

[tool.setuptools]
include-package-data = false

[tool.mypy]
# plugins = ["mypy_django_plugin.main"]
#strict_optional = true
python_version = "3.11"
disallow_untyped_defs = true
#show_error_codes = true
#no_implicit_optional = true
#warn_return_any = true
#warn_unused_ignores = true
exclude = ["scripts", "docs", "test"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests"]
